move AbstractExpectedSegment to generic package
