+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/predicate/ExpectedColumnSegment.java
@@ -20,6 +20,7 @@ package org.apache.shardingsphere.sql.parser.integrate.jaxb.predicate;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.sql.parser.core.constant.QuoteCharacter;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.generic.AbstractExpectedSegment;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.projection.ExpectedTableSegment;
 
 import javax.xml.bind.annotation.XmlAccessType;
